+++ packages/adyen-integration/src/adyenv2/create-adyenv2-payment-strategy.ts
@@ -1,12 +1,12 @@
 import { getScriptLoader, getStylesheetLoader } from '@bigcommerce/script-loader';
 
+import { AdyenV2ScriptLoader } from '@bigcommerce/checkout-sdk/adyen-utils';
 import {
     PaymentStrategyFactory,
     toResolvableModule,
 } from '@bigcommerce/checkout-sdk/payment-integration-api';
 
 import AdyenV2PaymentStrategy from './adyenv2-payment-strategy';
-import AdyenV2ScriptLoader from './adyenv2-script-loader';
 
 const createAdyenV2PaymentStrategy: PaymentStrategyFactory<AdyenV2PaymentStrategy> = (
     paymentIntegrationService,
