+++ packages/adyen-integration/src/index.ts
@@ -1,12 +1,2 @@
 export { default as createAdyenV2PaymentStrategy } from './adyenv2/create-adyenv2-payment-strategy';
 export { default as createAdyenV3PaymentStrategy } from './adyenv3/create-adyenv3-payment-strategy';
-export {
-    default as AdyenV2PaymentInitializeOptions,
-    WithAdyenV2PaymentInitializeOptions,
-} from './adyenv2/adyenv2-initialize-options';
-export {
-    default as AdyenV3PaymentInitializeOptions,
-    WithAdyenV3PaymentInitializeOptions,
-} from './adyenv3/adyenv3-initialize-options';
-export { default as AdyenV2ScriptLoader } from './adyenv2/adyenv2-script-loader';
-export { default as AdyenV3ScriptLoader } from './adyenv3/adyenv3-script-loader';
