+++ packages/adyen-utils/.eslintrc.json
@@ -1,0 +1,38 @@
+{
+    "extends": ["../../.eslintrc.json"],
+    "ignorePatterns": ["!**/*"],
+    "overrides": [
+        {
+            "files": ["*.ts", "*.tsx", "*.js", "*.jsx"],
+            "rules": {
+                "@typescript-eslint/naming-convention": "off",
+                "@typescript-eslint/consistent-type-assertions": "off",
+                "@typescript-eslint/no-unsafe-assignment": "off",
+                "@typescript-eslint/no-unsafe-argument": "off",
+                "@typescript-eslint/no-floating-promises": "off",
+                "@typescript-eslint/no-unsafe-call": "off",
+                "@typescript-eslint/no-explicit-any": "off",
+                "@typescript-eslint/no-unsafe-member-access": "off",
+                "@typescript-eslint/no-unsafe-return": "off"
+            }
+        },
+        {
+            "files": ["*.spec.ts"],
+            "rules": {
+                "jest/valid-expect": "off",
+                "jest/no-if": "off",
+                "@typescript-eslint/await-thenable": "off",
+                "@typescript-eslint/no-unsafe-assignment": "off",
+                "jest/no-conditional-expect": "off",
+                "jest/no-test-return-statement": "off",
+                "@typescript-eslint/no-shadow": "off"
+            }
+        },
+        {
+            "files": ["*.mock.ts"],
+            "rules": {
+                "@typescript-eslint/no-explicit-any": "off"
+            }
+        }
+    ]
+}
