+++ packages/adyen-utils/README.md
@@ -1,0 +1,11 @@
+# adyen-utils
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test adyen-utils` to execute the unit tests via [Jest](https://jestjs.io).
+
+## Running lint
+
+Run `nx lint adyen-utils` to execute the lint via [ESLint](https://eslint.org/).
