+++ packages/adyen-utils/jest.config.js
@@ -1,0 +1,12 @@
+module.exports = {
+    displayName: 'adyen-utils',
+    preset: '../../jest.preset.js',
+    globals: {
+        'ts-jest': {
+            tsconfig: '<rootDir>/tsconfig.spec.json',
+            diagnostics: false,
+        },
+    },
+    setupFilesAfterEnv: ['../../jest-setup.js'],
+    coverageDirectory: '../../coverage/packages/adyen-utils',
+};
