+++ packages/adyen-utils/src/adyenv3/adyenv3-initialize-options.ts
@@ -3,15 +3,13 @@ import { Omit } from '@bigcommerce/checkout-sdk/payment-integration-api';
 import {
     AdyenAdditionalActionOptions,
     AdyenCreditCardComponentOptions,
-    AdyenIdealComponentOptions,
-    AdyenThreeDS2Options,
-    AdyenV2ValidationState,
-} from './adyenv2';
+    AdyenValidationState,
+} from '../types';
 
 /**
- * A set of options that are required to initialize the AdyenV2 payment method.
+ * A set of options that are required to initialize the Adyenv3 payment method.
  *
- * Once AdyenV2 payment is initialized, credit card form fields, provided by the
+ * Once Adyenv3 payment is initialized, credit card form fields, provided by the
  * payment provider as IFrames, will be inserted into the current page. These
  * options provide a location and styling for each of the form fields.
  *
@@ -25,8 +23,8 @@ import {
  *
  * ```js
  * service.initializePayment({
- *     methodId: 'adyenv2',
- *     adyenv2: {
+ *     methodId: 'adyenv3',
+ *     adyenv3: {
  *         containerId: 'container',
  *         additionalActionOptions: {
  *             containerId: 'additional-action-container',
@@ -40,8 +38,8 @@ import {
  *
  * ```js
  * service.initializePayment({
- *     methodId: 'adyenv2',
- *     adyenv2: {
+ *     methodId: 'adyenv3',
+ *     adyenv3: {
  *         containerId: 'container',
  *         additionalActionOptions: {
  *             containerId: 'additional-action-container',
@@ -59,28 +57,16 @@ import {
  *             scheme: {
  *                 hasHolderName: true,
  *             },
- *             bcmc: {
- *                 hasHolderName: true,
- *             },
- *             ideal: {
- *                 showImage: true,
- *             },
  *         },
  *     },
  * });
  * ```
  */
-export default interface AdyenV2PaymentInitializeOptions {
+export default interface AdyenV3PaymentInitializeOptions {
     /**
      * The location to insert the Adyen component.
      */
     containerId: string;
-
-    /**
-     * @deprecated The location to insert the Adyen 3DS V2 component.
-     * Use additionalActionOptions instead as this property will be removed in the future
-     */
-    threeDS2ContainerId: string;
 
     /**
      * The location to insert the Adyen custom card component
@@ -93,12 +79,6 @@ export default interface AdyenV2PaymentInitializeOptions {
     hasVaultedInstruments?: boolean;
 
     /**
-     * @deprecated
-     * Use additionalActionOptions instead as this property will be removed in the future
-     */
-    threeDS2Options?: AdyenThreeDS2Options;
-
-    /**
      * A set of options that are required to initialize additional payment actions.
      */
     additionalActionOptions: AdyenAdditionalActionOptions;
@@ -106,17 +86,17 @@ export default interface AdyenV2PaymentInitializeOptions {
     /**
      * Optional. Overwriting the default options
      */
-    options?: Omit<AdyenCreditCardComponentOptions, 'onChange'> | AdyenIdealComponentOptions;
+    options?: Omit<AdyenCreditCardComponentOptions, 'onChange'>;
 
     shouldShowNumberField?: boolean;
 
-    validateCardFields(validateState: AdyenV2ValidationState): void;
+    validateCardFields(validateState: AdyenValidationState): void;
 }
 
-export interface WithAdyenV2PaymentInitializeOptions {
+export interface WithAdyenV3PaymentInitializeOptions {
     /**
      * The options that are required to initialize the Apple Pay payment
      * method. They can be omitted unless you need to support Apple Pay.
      */
-    adyenv2?: AdyenV2PaymentInitializeOptions;
+    adyenv3?: AdyenV3PaymentInitializeOptions;
 }
